v2v: chmod original OVA file if running as root (RHBZ#1430680).
In the case where we are going to read the disk directly from the OVA
file (partial = true), we will create a qcow2 image backed by the OVA.
If running as root, libvirt will run qemu as a non-root user (because
of no qemu:///session support for root, which is a libvirt bug).  qemu
will not be able to read the backing file and thus will fail.

To work around this problem, chmod the backing (OVA) file.

This isn't a complete fix since (eg) the OVA might be located in a
directory which isn't accessible by root, but there's only so much we
can do here until libvirt is fixed.

Thanks: Kun Wei.
(cherry picked from commit 542eb560bd8c7b24d4628837730d636713d02f56)